Extract ensureBadgeVisibilityColumns into shared util to eliminate duplication
The function was defined identically in both userController and badgeController.
It now lives in utils/badgeVisibilityUtils.js and is imported where needed.
