+++ src/controllers/badgeController.js
@@ -1,4 +1,5 @@
 const { pool } = require("../config/database");
+const { ensureBadgeVisibilityColumns } = require("../utils/badgeVisibilityUtils");
 
 /**
  * @description Get all badges grouped by category
@@ -56,15 +57,6 @@ const refreshBadgeViews = async (clientOrPool) => {
       }
     }
   }
-};
-
-const ensureBadgeVisibilityColumns = async (clientOrPool = pool) => {
-  await clientOrPool.query(
-    `ALTER TABLE users
-     ADD COLUMN IF NOT EXISTS hide_badges BOOLEAN DEFAULT FALSE,
-     ADD COLUMN IF NOT EXISTS hidden_badge_ids INTEGER[] DEFAULT '{}'::INTEGER[],
-     ADD COLUMN IF NOT EXISTS hidden_award_ids INTEGER[] DEFAULT '{}'::INTEGER[]`,
-  );
 };
 
 // ============================================================================
